add some casts to allow compilation for arm when some...
..."float optimization" is used (otherwise compilation breaks!)

+++ src/ai/default/move.cpp
@@ -477,7 +477,7 @@ std::pair<map_location,map_location> ai_default::choose_move(std::vector<target>
 			raise_user_interact();
 
 			const move_cost_calculator calc(u->second, map_, units_, enemy_dstsrc);
-			const double locStopValue = std::min(best_target->value / best_rating, 100.0);
+			const double locStopValue = std::min(best_target->value / best_rating, (double) 100.0);
 			pathfind::plain_route cur_route = a_star_search(u->first, best_target->loc, locStopValue, &calc, map_.w(), map_.h());
 
 			if(cur_route.steps.empty()) {
